verified organizations connections refactored
