+++ api-js/src/verified-organizations/objects/index.js
@@ -1,1 +1,2 @@
 export * from './verified-organization'
+export * from './verified-organization-connection'
